pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantriver"
version = "0.0.1"
description = "A public event-driven trading engine with backtest, paper, and live runners."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "QuantRiver" }
]
keywords = ["trading", "backtesting", "event-driven", "market-data", "python"]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.0",
  "pyarrow>=14.0",
  "requests>=2.31",
  "tzdata>=2024.1",
  "websocket-client>=1.8",
]

[project.optional-dependencies]
gates = [
  "lightgbm>=4.3",
]

[tool.setuptools.packages.find]
include = ["adapters*", "core*", "runners*"]
